Better packaging of Python modules.
