+++ src/main/python/aut/__init__.py
@@ -1,0 +1,5 @@
+from aut.common import WebArchive
+from aut.udfs import extract_domain
+
+__all__ = ['WebArchive', 'extract_domain']
+
